+++ Editor/Playa/Renderer/SerializedFieldRendererUIToolkit.cs
@@ -35,7 +35,7 @@ namespace SaintsField.Editor.Playa.Renderer
 
 
             // About letting SaintsPropertyDrawer fallback:
-            // SaintsPropertyDrawer relys on PropertyField to fallback. Directly hi-jiacking the drawer with SaintsPropertyDrawer
+            // SaintsPropertyDrawer relays on PropertyField to fallback. Directly hi-jacking the drawer with SaintsPropertyDrawer
             // the workflow will still get into the PropertyField flow, then SaintsField will fail to decide when the
             // fallback should stop.
             PropertyAttribute[] allAttributes = ReflectCache.GetCustomAttributes<PropertyAttribute>(FieldWithInfo.FieldInfo);
@@ -112,7 +112,7 @@ namespace SaintsField.Editor.Playa.Renderer
                     this,
                     null
                 );
-                return (MergeWithDec(r, allAttributes), false);
+                return (UIToolkitCache.MergeWithDec(r, allAttributes), false);
             }
 
             // Nah... This didn't handle for mis-ordered case
@@ -149,7 +149,7 @@ namespace SaintsField.Editor.Playa.Renderer
             if (!useImGui)
             {
                 VisualElement r = propertyDrawer.CreatePropertyGUI(FieldWithInfo.SerializedProperty);
-                return (MergeWithDec(r, allAttributes), false);
+                return (UIToolkitCache.MergeWithDec(r, allAttributes), false);
             }
 
             // SaintsPropertyDrawer won't have pure IMGUI one. Let Unity handle it.
@@ -203,108 +203,6 @@ namespace SaintsField.Editor.Playa.Renderer
             // imGuiContainer.AddToClassList(IMGUILabelHelper.ClassName);
             //
             // return (imGuiContainer, false);
-        }
-
-        private static VisualElement MergeWithDec(VisualElement result, PropertyAttribute[] allAttributes)
-        {
-            VisualElement dec = DrawDecorator(allAttributes);
-            if(dec == null)
-            {
-                return result;
-            }
-
-            VisualElement container = new VisualElement();
-            container.Add(dec);
-            container.Add(result);
-            return container;
-        }
-
-
-        private static VisualElement DrawDecorator(IEnumerable<PropertyAttribute> allPropAttributes)
-        {
-            IReadOnlyDictionary<Type, IReadOnlyList<SaintsPropertyDrawer.PropertyDrawerInfo>> propertyAttributeToDecoratorDrawers = SaintsPropertyDrawer.EnsureAndGetTypeToDrawers().attrToDecoratorDrawers;
-            // this can be multiple, should not be a dictionary
-            IEnumerable<(PropertyAttribute propAttribute, Type)> decDrawers =
-                allPropAttributes
-                    .Select(propAttribute => (propAttribute, SaintsPropertyDrawer.PropertyGetDecoratorDrawer(propAttribute.GetType())))
-                    .Where(each => each.Item2 != null);
-            // new List<(PropertyAttribute, Type drawerType)>();
-            // if (decDrawers.Length == 0)
-            // {
-            //     return null;
-            // }
-
-            VisualElement result = new VisualElement();
-            result.AddToClassList("unity-decorator-drawers-container");
-            // this.m_DecoratorDrawersContainer = new VisualElement();
-            // this.m_DecoratorDrawersContainer.AddToClassList(PropertyField.decoratorDrawersContainerClassName);
-
-            bool hasAny = false;
-            foreach ((PropertyAttribute propAttribute, Type drawerType) in decDrawers)
-            {
-                hasAny = true;
-                DecoratorDrawer decorator = (DecoratorDrawer)Activator.CreateInstance(drawerType);
-                FieldInfo mAttributeField = drawerType.GetField("m_Attribute", BindingFlags.NonPublic | BindingFlags.Instance);
-                if (mAttributeField != null)
-                {
-                    mAttributeField.SetValue(decorator, propAttribute);
-                }
-
-                VisualElement ve =
-#if UNITY_2022_3_OR_NEWER
-                        decorator.CreatePropertyGUI()
-#else
-                        null
-#endif
-                    ;
-
-                // ReSharper disable once ConditionIsAlwaysTrueOrFalse
-                if (ve == null)
-                {
-                    ve = new IMGUIContainer(() =>
-                    {
-                        Rect position = new Rect
-                        {
-                            height = decorator.GetHeight(),
-                            width = result.resolvedStyle.width,
-                        };
-                        decorator.OnGUI(position);
-                        // ReSharper disable once PossibleNullReferenceException
-                        // ReSharper disable once AccessToModifiedClosure
-                        ve.style.height = position.height;
-                    });
-                    ve.style.height = decorator.GetHeight();
-                }
-                result.Add(ve);
-            }
-
-            // foreach (DecoratorDrawer decoratorDrawer in decoratorDrawers)
-            // {
-            //     DecoratorDrawer decorator = decoratorDrawer;
-            //     VisualElement ve = decorator.CreatePropertyGUI();
-            //     if (ve == null)
-            //     {
-            //         ve = (VisualElement) new IMGUIContainer((Action) (() =>
-            //         {
-            //             Rect position = new Rect();
-            //             position.height = decorator.GetHeight();
-            //             position.width = this.resolvedStyle.width;
-            //             decorator.OnGUI(position);
-            //             ve.style.height = (StyleLength) position.height;
-            //         }));
-            //         ve.style.height = (StyleLength) decorator.GetHeight();
-            //     }
-            //     this.m_DecoratorDrawersContainer.Add(ve);
-            // }
-
-            return hasAny? result: null;
-
-            // // PropertyHandler propertyHandle = UnityEditor.ScriptAttributeUtility.GetHandler(FieldWithInfo.SerializedProperty);
-            // PropertyHandler propertyHandle = (PropertyHandler)typeof(UnityEditor.Editor).Assembly
-            //     .GetType("UnityEditor.ScriptAttributeUtility")
-            //     .GetMethod("GetHandler", BindingFlags.NonPublic | BindingFlags.Static)
-            //     .Invoke(null, new object[] { FieldWithInfo.SerializedProperty });
-
         }
 
         public AbsRenderer MakeRenderer(SerializedObject serializedObject, SaintsFieldWithInfo fieldWithInfo)