+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.3.1 ##
+
+1.  Fix `SaintsEditor` can not find a correct typed drawer on first drawing after a domain reload
+2.  `SaintsDictionary` now can use the new fallback system, which means if you have a struct/class typed key/value, you can use `Layout`, `Button` etc. in it without add a `SaintsRow` decorator
+
 ## 4.3.0 ##
 
 1.  **Breaking Changes**: `SepTitle` arguments now aligned with `Separator`: `string title`, `EColor color`, `EAlign eAlign`, `int space`